refactor :package: move OfferWizardStepView to separate file
